Cleanup: Missing documentation + consistency fixes
